add missing test case for exercise 2 - task 1 and task 2
Current tests for Task 1 and Task 2 of exercise 2 don't cover the case where the highest number in the list is at index 0. In some implementations, this can lead to false positives in test, where, for example, for the input [3, 1, 2] the result (3, 3) is considered as a correct even though it is not.